client gfx refactor (1): collect gfx code in module
