+++ alvr/client_core/Cargo.toml
@@ -22,8 +22,10 @@ alvr_sockets.workspace = true
 
 app_dirs2 = "2"
 bincode = "1"
+glow = "0.13"
 glyph_brush_layout = "0.2"
 jni = "0.21"
+khronos-egl = { version = "6", features = ["dynamic"] }
 local-ip-address = "0.6"
 mdns-sd = "0.11"
 rand = "0.8"
@@ -33,13 +35,10 @@ whoami = "1"
 
 [target.'cfg(target_os = "android")'.dependencies]
 android_logger = "0.13"
-ndk = { version = "0.9", features = [
-    "api-level-26",
-    "media",
-] }
+ndk = { version = "0.9", features = ["api-level-26", "media"] }
 ndk-context = "0.1"
 ndk-sys = "0.6"
-oboe = "0.6" # todo: remove once AudioThread shutdown crash is fixed
+oboe = "0.6"                                                    # todo: remove once AudioThread shutdown crash is fixed
 
 [target.'cfg(not(target_os = "android"))'.dependencies]
 env_logger = "0.11"
