+++ alvr/client_openxr/Cargo.toml
@@ -15,8 +15,6 @@ alvr_client_core.workspace = true
 alvr_packets.workspace = true
 alvr_session.workspace = true
 
-glow = "0.13"
-khronos-egl = { version = "6", features = ["dynamic"] }
 openxr = { git = "https://github.com/alvr-org/openxrs", rev = "f1f47ff" }
 # openxr = { path = "../../../../openxrs/openxr" }
